Refactor build workflow to retrieve version information and generate release package
Updates the build workflow to retrieve the version number from the Git tag and include it in the release package name. Adds functionality to retrieve the tag message and changes since the last release. Improves the release package creation process by creating a directory and copying necessary files. Updates the release action to include the version number, tag message, and changes in the release notes.